Allow openstack backend to connect to the vm through a proxy (#19)
* Allow openstack backend to connect to the vm through a proxy

When access to the VM must go through a proxy, the proxy configuration
should be specified using the proxy key. Additionally, a mapping between
he destination IP address and the proxy port must be defined.

To configure the IP-port mapping, use the cidr-port-rel key. This
requires a list of entries in the format "CIDR:initial-port". Based on
this list, the system will assign a port by matching the allocated IP
with the corresponding CIDR and applying the specified initial port.

* use environments to test spread-plus in arm and amd

* Fix backend name

* split openstack in different jobs

* skip cleanup at the end to be able to run

* fix ps7 image names