Skip condition
A skip section determines whether a task should be skipped before it
begins execution

The skip section can only be defined at the task level. They are
evaluated before the task preparation phase

The SKIP command is a simple mechanism used to indicate that a task
should not run under certain conditions.